Splitted model into several files in lib/model/
